backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "CPD Events Platform Backend"
authors = [
    {name = "Rutvik Solanki",email = "rutvikrj26@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    # Django core
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.15.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-cors-headers (>=4.3.0,<5.0.0)",
    "drf-yasg (>=1.21.0,<2.0.0)",
    "drf-nested-routers (>=0.94.0,<1.0.0)",
    
    # Database & Caching
    "psycopg[binary] (>=3.1.0,<4.0.0)",
    "redis (>=5.0.0,<6.0.0)",
    
    # Task Queue (GCP Cloud Tasks)
    "google-cloud-tasks (>=2.16.0,<3.0.0)",
    
    # Payments
    "stripe (>=8.0.0,<9.0.0)",
    
    # Cloud Storage
    "django-storages[google] (>=1.14.0,<2.0.0)",
    "google-cloud-storage (>=2.14.0,<3.0.0)",
    
    # Security & Encryption
    "cryptography (>=46.0.3,<47.0.0)",
    
    # PDF Generation
    "reportlab (>=4.1.0,<5.0.0)",
    "Pillow (>=10.2.0,<11.0.0)",
    
    # HTTP Requests
    "requests (>=2.31.0,<3.0.0)",
    
    # Utilities
    "pytz (>=2025.2,<2026.0)",
    "python-dateutil (>=2.8.0,<3.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-django (>=4.7.0,<5.0.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
    "factory-boy (>=3.3.0,<4.0.0)",
    "faker (>=22.0.0,<23.0.0)",
    "black (>=24.0.0,<25.0.0)",
    "ruff (>=0.2.0,<1.0.0)",
    "mypy (>=1.8.0,<2.0.0)",
    "django-stubs (>=4.2.0,<5.0.0)",
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
testpaths = ["src"]

[tool.black]
line-length = 128
target-version = ['py313']
include = '\.pyi?$'
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 128
src_paths = ["src"]
force_single_line = false
combine_as_imports = true
balanced_wrapping = true

[tool.ruff]
line-length = 128
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.pylint.messages_control]
disable = [
    "too-few-public-methods",
    "wrong-import-order",
    "redefined-outer-name",
    "unused-argument",
    "too-many-instance-attributes",
    "duplicate-code",
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-many-branches",
    "too-many-statements",
    "too-many-locals",
    "broad-exception-caught",
    "import-outside-toplevel",
    "fixme",
    "missing-module-docstring",      # C0114
    "missing-class-docstring",       # C0115
    "missing-function-docstring",    # C0116
    "logging-fstring-interpolation", # W1203
]

[tool.pylint.format]
max-line-length = 128

[tool.pyright]
include = ["src"]
pythonVersion = "3.13"
typeCheckingMode = "standard"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnknownVariableType = true
reportUnusedImport = true
